+++ src/main/java/frostvein/sampires/remakepire/abilities/tome/UncannyDirectionTomeAbility.java
@@ -14,9 +14,11 @@ public class UncannyDirectionTomeAbility extends TomeAbility {
         if (!this.canUse(player)) {
             this.sendCannotUseMessage(player, "Only humans can use tome abilities!");
             return false;
+
         } else {
             final double townCenterX = this.plugin.getConfigManager().getOakhurstTownCenterX();
             final double townCenterZ = this.plugin.getConfigManager().getOakhurstTownCenterZ();
+
             (new BukkitRunnable() {
                 int ticksRemaining = 140;
 
@@ -26,10 +28,12 @@ public class UncannyDirectionTomeAbility extends TomeAbility {
                         double deltaX = townCenterX - currentLocation.getX();
                         double deltaZ = townCenterZ - currentLocation.getZ();
                         double distance = Math.sqrt(deltaX * deltaX + deltaZ * deltaZ);
+
                         String direction = UncannyDirectionTomeAbility.this.getRelativeDirection(deltaX, deltaZ, currentLocation.getYaw());
                         String actionBarMessage = String.format("§6Town Center: §f %s §7(§f%.0f blocks§7)", direction, distance);
                         UncannyDirectionTomeAbility.this.plugin.getSessionManager().sendActionBar(player, actionBarMessage);
                         this.ticksRemaining -= 4;
+
                     } else {
                         this.cancel();
                     }
@@ -44,31 +48,31 @@ public class UncannyDirectionTomeAbility extends TomeAbility {
     private String getRelativeDirection(double deltaX, double deltaZ, float playerYaw) {
         double targetAngle = Math.atan2(deltaX, -deltaZ);
         double targetDegrees = Math.toDegrees(targetAngle);
-        if (targetDegrees < (double)0.0F) {
-            targetDegrees += (double)360.0F;
+        if (targetDegrees < 0.0) {
+            targetDegrees += 360.0;
         }
 
-        double playerFacing = (double)((playerYaw + 180.0F) % 360.0F);
-        if (playerFacing < (double)0.0F) {
-            playerFacing += (double)360.0F;
+        double playerFacing = (playerYaw + 180.0F) % 360.0F;
+        if (playerFacing < 0.0) {
+            playerFacing += 360.0;
         }
 
-        double relativeAngle = (targetDegrees - playerFacing + (double)360.0F) % (double)360.0F;
-        if (!(relativeAngle >= (double)337.5F) && !(relativeAngle < (double)22.5F)) {
-            if (relativeAngle >= (double)22.5F && relativeAngle < (double)67.5F) {
+        double relativeAngle = (targetDegrees - playerFacing + 360.0) % 360.0;
+        if (!(relativeAngle >= 337.5) && !(relativeAngle < 22.5)) {
+            if (relativeAngle >= 22.5 && relativeAngle < 67.5) {
                 return "\ue00b";
-            } else if (relativeAngle >= (double)67.5F && relativeAngle < (double)112.5F) {
+            } else if (relativeAngle >= 67.5 && relativeAngle < 112.5) {
                 return "\ue00c";
-            } else if (relativeAngle >= (double)112.5F && relativeAngle < (double)157.5F) {
+            } else if (relativeAngle >= 112.5 && relativeAngle < 157.5) {
                 return "\ue00d";
-            } else if (relativeAngle >= (double)157.5F && relativeAngle < (double)202.5F) {
+            } else if (relativeAngle >= 157.5 && relativeAngle < 202.5) {
                 return "\ue00e";
-            } else if (relativeAngle >= (double)202.5F && relativeAngle < (double)247.5F) {
+            } else if (relativeAngle >= 202.5 && relativeAngle < 247.5) {
                 return "\ue00f";
-            } else if (relativeAngle >= (double)247.5F && relativeAngle < (double)292.5F) {
+            } else if (relativeAngle >= 247.5 && relativeAngle < 292.5) {
                 return "\ue010";
             } else {
-                return relativeAngle >= (double)292.5F && relativeAngle < (double)337.5F ? "\ue011" : "\ue00a";
+                return relativeAngle >= 292.5 && relativeAngle < 337.5 ? "\ue011" : "\ue00a";
             }
         } else {
             return "\ue00a";
