+++ src/main/java/frostvein/sampires/remakepire/managers/BeaconMajorityManager.java
@@ -61,7 +61,7 @@ public class BeaconMajorityManager {
     private void applyBonusToHumans(int bonusHearts) {
         this.currentHumanBonus = bonusHearts;
         this.currentVampireBonus = 0;
-        double healthBonus = (double)bonusHearts * 2.0;
+        double healthBonus = bonusHearts * 2.0;
 
         for(Player player : Bukkit.getOnlinePlayers()) {
             if (this.vampireManager.isHuman(player)) {
@@ -75,7 +75,7 @@ public class BeaconMajorityManager {
     private void applyBonusToVampires(int bonusHearts) {
         this.currentVampireBonus = bonusHearts;
         this.currentHumanBonus = 0;
-        double healthBonus = (double)bonusHearts * 2.0;
+        double healthBonus = bonusHearts * 2.0;
 
         for(Player player : Bukkit.getOnlinePlayers()) {
             if (this.vampireManager.isVampire(player)) {
@@ -130,7 +130,7 @@ public class BeaconMajorityManager {
             double currentHealth = player.getHealth();
             double maxHealth = player.getMaxHealth();
 
-            if (currentHealth > (double)0.0F && currentHealth >= maxHealth - healthBonus) {
+            if (currentHealth > 0 && currentHealth >= maxHealth - healthBonus) {
                 player.setHealth(player.getMaxHealth());
             }
 
@@ -160,11 +160,12 @@ public class BeaconMajorityManager {
     public void applyBonusesToPlayer(Player player) {
         if (this.plugin.getSessionManager().isSessionActive()) {
             if (this.vampireManager.isVampire(player) && this.currentVampireBonus > 0) {
-                double healthBonus = (double)this.currentVampireBonus * (double)2.0F;
+                double healthBonus = this.currentVampireBonus * 2.0;
                 this.applyHealthModifier(player, healthBonus, VAMPIRE_MAJORITY_HEALTH_UUID, "Beacon Majority (Vampire)");
+
             } else if (this.vampireManager.isHuman(player)) {
                 if (this.currentHumanBonus > 0) {
-                    double healthBonus = (double)this.currentHumanBonus * (double)2.0F;
+                    double healthBonus = this.currentHumanBonus * 2.0;
                     this.applyHealthModifier(player, healthBonus, HUMAN_MAJORITY_HEALTH_UUID, "Beacon Majority (Human)");
                 }
 
@@ -183,10 +184,12 @@ public class BeaconMajorityManager {
     private void applyDeathPenalty(Player player) {
         if (this.vampireManager.isHuman(player)) {
             int deathCount = this.getPlayerDeathCount(player);
+
             if (deathCount > 0) {
-                double healthPenalty = -((double)deathCount * 2.0);
+                double healthPenalty = -(deathCount * 2.0);
                 this.applyHealthModifier(player, healthPenalty, DEATH_PENALTY_HEALTH_UUID, "Death Penalty");
                 this.plugin.getLogger().fine("Applied -" + deathCount + " hearts death penalty to " + player.getName());
+
             } else {
                 this.removeHealthModifier(player, DEATH_PENALTY_HEALTH_UUID);
             }
